# Conflicts:
#	CardBox.xcodeproj/project.pbxproj
#	CardBox/Views/GameRunnerView.swift
#	CardBox/Views/PlayerView.swift

+++ CardBox/Views/PlayerHandView.swift
@@ -25,13 +25,14 @@ struct PlayerHandView: View {
     }
 
     var body: some View {
-        HStack(spacing: spacing) {
+        HStack(spacing: CGFloat(spacing)) {
             ForEach(playerHandViewModel.getCards()) { card in
                 let cardViewModel = CardViewModel(card: card, isFaceUp: true)
                 CardView(cardViewModel: cardViewModel)
                     .onTapGesture {
                         print("tap card")
-                        playerViewModel.tapCard(card: card, cardViewModel: cardViewModel, gameRunner: gameRunnerViewModel)
+                        playerViewModel
+                            .tapCard(card: card, cardViewModel: cardViewModel, gameRunner: gameRunnerViewModel)
                     }
                     .gesture(
                         DragGesture(minimumDistance: 0.0)
